Moved PostFormInfo from Helpers to RestClient namespace.
