+++ source/restclient.cc
@@ -82,7 +82,7 @@ RestClient::Response RestClient::post(const std::string& url,
  * @return response struct
  */
 RestClient::Response RestClient::postForm(const std::string& url,
-                    const RestClient::Helpers::PostFormInfo& data) {
+                    const PostFormInfo& data) {
   RestClient::Response ret;
   RestClient::Connection *conn = new RestClient::Connection("");
   ret = conn->postForm(url, data);
@@ -138,4 +138,52 @@ RestClient::Response RestClient::head(const std::string& url) {
   ret = conn->head(url);
   delete conn;
   return ret;
+}
+
+/**
+ * @brief PostFormInfo constructor
+ */
+RestClient::PostFormInfo::PostFormInfo()
+        : formPtr(NULL), lastFormPtr(NULL) {
+}
+
+/**
+ * @brief PostFormInfo destructor
+ */
+RestClient::PostFormInfo::~PostFormInfo() {
+  // cleanup the formpost chain
+  if (this->formPtr) {
+    curl_formfree(this->formPtr);
+    this->formPtr = NULL;
+    this->lastFormPtr = NULL;
+  }
+}
+
+/**
+ * @brief set the name and the value of the HTML "file" form's input
+ *
+ * @param fieldName name of the "file" input
+ * @param fieldValue path to the file to upload
+ */
+void RestClient::PostFormInfo::addFormFile(
+          const std::string& fieldName, const std::string& fieldValue) {
+  curl_formadd(&this->formPtr, &this->lastFormPtr,
+               CURLFORM_COPYNAME, fieldName.c_str(),
+               CURLFORM_FILE, fieldValue.c_str(),
+               CURLFORM_END);
+}
+
+/**
+ * @brief set the name and the value of an HTML form's input 
+ * (other than "file" like "text", "hidden" or "submit")
+ *
+ * @param fieldName name of the input element
+ * @param fieldValue value to be assigned to the input element
+ */
+void RestClient::PostFormInfo::addFormContent(
+          const std::string& fieldName, const std::string& fieldValue) {
+  curl_formadd(&this->formPtr, &this->lastFormPtr,
+               CURLFORM_COPYNAME, fieldName.c_str(),
+               CURLFORM_COPYCONTENTS, fieldValue.c_str(),
+               CURLFORM_END);
 }
